+++ crates/ruma-federation-api/CHANGELOG.md
@@ -11,6 +11,7 @@ Improvements:
 
 * Move `knock` module out of `unstable-pre-spec`
   * `knock:::send_knock::v1::Request` requires a PDU instead of the `knock_event`
+* Move cross-signing properties of `keys::get_keys::v1::Response` out of `unstable-pre-spec`
 
 # 0.3.1